fix errors

+++ control/MSingleEdit.h
@@ -45,7 +45,7 @@ public:
 		pDC->SetBkMode(TRANSPARENT);
 		pDC->SetTextColor(*m_color_text);
 		CSize sz;
-		::GetTextExtentPoint32(pDC->GetSafeHdc(), m_text.data(), m_text.length(), &sz);
+		::GetTextExtentPoint32(pDC->GetSafeHdc(), m_text.data(), static_cast<int>(m_text.length()), &sz);
 
 		pDC->TextOut(rect.left, (rect.Height() - sz.cy) / 2 + rect.top, m_text.data());
